refactored Oakparticle, and added Oak class
