+++ libraries/OakParticle/Oak.cpp
@@ -1,0 +1,10 @@
+
+#include "Oak.h"
+
+extern void rebootToUser();
+extern void rebootToConfig();
+
+void OakClass::rebootToUser(void) { rebootToUser(); }
+void OakClass::rebootToConfig(void) { rebootToConfig(); }
+
+OakClass Oak;
