+++ libraries/OakParticle/Oak.h
@@ -1,0 +1,12 @@
+#ifndef Oak_h
+#define Oak_h
+
+class OakClass {
+public:
+	void rebootToUser(void);
+	void rebootToConfig(void);
+};
+
+extern OakClass Oak;
+
+#endif // Oak_h
