+++ libraries/OakParticle/OakParticle.h
@@ -1,6 +1,6 @@
 /**
  ******************************************************************************
- * @file    spark_wiring_cloud.h
+ * @file    OakParticle.h
  * @author  Satish Nair, Zachary Crockett, Matthew McGowan
  ******************************************************************************
   Copyright (c) 2013-2015 Particle Industries, Inc.  All rights reserved.
@@ -23,7 +23,8 @@
 #ifndef OakParticle_h
 #define OakParticle_h
 
-#include "particle_core.h"
+#include <cstring>
+#include "particle.h"
 
 #ifdef SPARK_NO_CLOUD
 #define CLOUD_FN(x,y) (y)
@@ -32,15 +33,9 @@
 #endif
 
 
-
 class CloudClass {
-
-
 public:
-
-    CloudClass(){
-        spark_initConfig();
-    }
+    CloudClass();
 
     static inline bool variable(const char* varKey, const int& var)
     {
@@ -83,20 +78,10 @@ public:
         return variable(varKey, (const char*)userVar, userVarType);
     }
 
-    template<typename T> static inline bool variable(const char *varKey, const typename T::varref userVar, const T& userVarType)
-    {
-        return CLOUD_FN(spark_variable(varKey, (const void*)userVar, T::value(), NULL), false);
-    }
+    template<typename T> static bool variable(const char *varKey, const typename T::varref userVar, const T& userVarType);
 
-    static inline bool variable(const char *varKey, const int32_t* userVar, const CloudVariableTypeInt& userVarType)
-    {
-        return CLOUD_FN(spark_variable(varKey, (const void*)userVar, CloudVariableTypeInt::value(), NULL), false);
-    }
-
-    static inline bool variable(const char *varKey, const uint32_t* userVar, const CloudVariableTypeInt& userVarType)
-    {
-        return CLOUD_FN(spark_variable(varKey, (const void*)userVar, CloudVariableTypeInt::value(), NULL), false);
-    }
+    static bool variable(const char *varKey, const int32_t* userVar, const CloudVariableTypeInt& userVarType);
+    static bool variable(const char *varKey, const uint32_t* userVar, const CloudVariableTypeInt& userVarType);
 
     // Return clear errors for common misuses of Particle.variable()
     template<typename T, std::size_t N>
@@ -107,13 +92,7 @@ public:
     }
 
     template<typename T>
-    static inline bool variable(const T *varKey, const String *userVar, const CloudVariableTypeString& userVarType)
-    {
-        spark_variable_t extra;
-        extra.size = sizeof(extra);
-        extra.update = update_string_variable;
-        return CLOUD_FN(spark_variable(varKey, userVar, CloudVariableTypeString::value(), &extra), false);
-    }
+    static bool variable(const T *varKey, const String *userVar, const CloudVariableTypeString& userVarType);
 
     template<typename T>
     static inline bool variable(const T *varKey, const String &userVar, const CloudVariableTypeString& userVarType)
@@ -122,92 +101,29 @@ public:
         return false;
     }
 
-    static bool function(const char *funcKey, user_function_int_str_t* func)
-    {
-        return CLOUD_FN(register_function(call_raw_user_function, (void*)func, funcKey), false);
-    }
-
-    static bool function(const char *funcKey, user_std_function_int_str_t func, void* reserved=NULL)
-    {
-#ifdef SPARK_NO_CLOUD
-        return false;
-#else
-        bool success = false;
-        if (func) // if the call-wrapper has wrapped a callable object
-        {
-            auto wrapper = new user_std_function_int_str_t(func);
-            if (wrapper) {
-                success = register_function(call_std_user_function, wrapper, funcKey);
-            }
-        }
-        return success;
-#endif
-    }
+    static bool function(const char *funcKey, user_function_int_str_t* func);
+    static bool function(const char *funcKey, user_std_function_int_str_t func, void* reserved=NULL);
 
     template <typename T>
-    static void function(const char *funcKey, int (T::*func)(String), T *instance) {
-      using namespace std::placeholders;
-      function(funcKey, std::bind(func, instance, _1));
-    }
+    static void function(const char *funcKey, int (T::*func)(String), T *instance);
 
-    bool publish(const char *eventName, Spark_Event_TypeDef eventType=PUBLIC)
-    {
-        return CLOUD_FN(spark_send_event(eventName, NULL, 60, eventType, NULL), false);
-    }
+    bool publish(const char *eventName, Spark_Event_TypeDef eventType=PUBLIC);
+    bool publish(const char *eventName, const char *eventData, Spark_Event_TypeDef eventType=PUBLIC);
+    bool publish(const char *eventName, const char *eventData, int ttl, Spark_Event_TypeDef eventType=PUBLIC);
 
-    bool publish(const char *eventName, const char *eventData, Spark_Event_TypeDef eventType=PUBLIC)
-    {
-        return CLOUD_FN(spark_send_event(eventName, eventData, 60, eventType, NULL), false);
-    }
-
-    bool publish(const char *eventName, const char *eventData, int ttl, Spark_Event_TypeDef eventType=PUBLIC)
-    {
-        return CLOUD_FN(spark_send_event(eventName, eventData, ttl, eventType, NULL), false);
-    }
-
-    bool subscribe(const char *eventName, EventHandler handler, Spark_Subscription_Scope_TypeDef scope=ALL_DEVICES)
-    {
-        return CLOUD_FN(spark_subscribe(eventName, handler, NULL, scope, NULL, NULL), false);
-    }
-
-    bool subscribe(const char *eventName, EventHandler handler, const char *deviceID)
-    {
-        return CLOUD_FN(spark_subscribe(eventName, handler, NULL, MY_DEVICES, deviceID, NULL), false);
-    }
-
-    bool subscribe(const char *eventName, wiring_event_handler_t handler, Spark_Subscription_Scope_TypeDef scope=ALL_DEVICES)
-    {
-        return subscribe_wiring(eventName, handler, scope);
-    }
-
-    bool subscribe(const char *eventName, wiring_event_handler_t handler, const char *deviceID)
-    {
-        return subscribe_wiring(eventName, handler, MY_DEVICES, deviceID);
-    }
+    bool subscribe(const char *eventName, EventHandler handler, Spark_Subscription_Scope_TypeDef scope=ALL_DEVICES);
+    bool subscribe(const char *eventName, EventHandler handler, const char *deviceID);
+    bool subscribe(const char *eventName, wiring_event_handler_t handler, Spark_Subscription_Scope_TypeDef scope=ALL_DEVICES);
+    bool subscribe(const char *eventName, wiring_event_handler_t handler, const char *deviceID);
 
     template <typename T>
-    bool subscribe(const char *eventName, void (T::*handler)(const char *, const char *), T *instance, Spark_Subscription_Scope_TypeDef scope=ALL_DEVICES)
-    {
-        using namespace std::placeholders;
-        return subscribe(eventName, std::bind(handler, instance, _1, _2), scope);
-    }
-
+    bool subscribe(const char *eventName, void (T::*handler)(const char *, const char *), T *instance, Spark_Subscription_Scope_TypeDef scope=ALL_DEVICES);
     template <typename T>
-    bool subscribe(const char *eventName, void (T::*handler)(const char *, const char *), T *instance, const char *deviceID)
-    {
-        using namespace std::placeholders;
-        return subscribe(eventName, std::bind(handler, instance, _1, _2), deviceID);
-    }
+    bool subscribe(const char *eventName, void (T::*handler)(const char *, const char *), T *instance, const char *deviceID);
 
-    void unsubscribe()
-    {
-        CLOUD_FN(remove_event_handlers(NULL), (void)0);
-    }
+    void unsubscribe();
 
-    bool syncTime(void)
-    {
-        return CLOUD_FN(send_time_request(),false);
-    }
+    bool syncTime(void);
 /*
     static void sleep(long seconds) __attribute__ ((deprecated("Please use System.sleep() instead.")))
     { SystemClass::sleep(seconds); }
@@ -216,13 +132,13 @@ public:
     static void sleep(uint16_t wakeUpPin, InterruptMode edgeTriggerMode, long seconds=0) __attribute__ ((deprecated("Please use System.sleep() instead.")))
     { SystemClass::sleep(wakeUpPin, edgeTriggerMode, seconds); }
 */
-    static void initialize(void) { return spark_initConfig(); }
-    static bool connected(void) { return spark_connected(); }
-    static bool disconnected(void) { return !connected(); }
-    static bool connect(void) { return spark_connect(); }
-    static void disconnect(void) { spark_disconnect(); }
-    static void process(void) { spark_process(); }
-    static String deviceID(void) { return spark_deviceID(); }
+    static void initialize(void);
+    static bool connected(void);
+    static bool disconnected(void);
+    static bool connect(void);
+    static void disconnect(void);
+    static void process(void);
+    static String deviceID(void);
 
 private:
 
@@ -232,30 +148,11 @@ private:
 
     //static void call_wiring_event_handler(const void* param, const char *event_name, const char *data);
 
-    bool subscribe_wiring(const char *eventName, wiring_event_handler_t handler, Spark_Subscription_Scope_TypeDef scope, const char *deviceID = NULL)
-    {
-#ifdef SPARK_NO_CLOUD
-        return false;
-#else
-        bool success = false;
-        if (handler) // if the call-wrapper has wrapped a callable object
-        {
-            auto wrapper = new wiring_event_handler_t(handler);
-            if (wrapper) {
-                success = spark_subscribe(eventName, (EventHandler)call_wiring_event_handler, wrapper, scope, deviceID, NULL);
-            }
-        }
-        return success;
-#endif
-    }
+    bool subscribe_wiring(const char *eventName, wiring_event_handler_t handler, Spark_Subscription_Scope_TypeDef scope, const char *deviceID = NULL);
 
-    static const void* update_string_variable(const char* name, Spark_Data_TypeDef type, const void* var, void* reserved)
-    {
-        const String* s = (const String*)var;
-        return s->c_str();
-    }
+    static const void* update_string_variable(const char* name, Spark_Data_TypeDef type, const void* var, void* reserved);
 };
 
 extern CloudClass Particle;
 
-#endif
+#endif // OakParticle_h
